+++ src/demo/Demo.vue
@@ -52,7 +52,12 @@
       <pre class="border"><code>
         &lt;slider-frame
           :slides-per-view="5"
-          :breakpoints="{sm: 1, md: 2, lg: 3, xl: 4}"
+          :breakpoints="{
+            sm: { slidesPerView: 1 },
+            md: { slidesPerView: 2 },
+            lg: { slidesPerView: 3 },
+            xl: { slidesPerView: 4 }
+          }"
         >
           &lt;div class="slider d-inline-flex" slot-scope="{ next, prev, scrollToSlide, activeIndex, canScrollNext, canScrollPrev }">
             &lt;button