+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # vue-dynamic-slider
 
+##### 0.3.0
+  - `ISSUE#1`: Add `pagination` slot prop that has number of pages and current page
+  - `ISSUE#2`: Add `scrollIncrement` prop that sets the number of slides that are scrolled at a time
+  - `ISSUE#3`: Add `slidesInView` slot prop that's an array of visible slide indices
+  - Handle touch devices better
+
 ##### 0.2.0
   - Clean up and optimize code
   - Add option to override scrolling function